Renamed `ecdsa` to `legacy`
The new name better aligns with keys not being part of signature scheme.
