+++ bitcoin_keys/src/legacy/serialized_public_key.rs
